refactor: complete Feature-Driven Architecture migration
- Remove empty components/ directory (all files migrated to features/ and shared/)
- Fix broken @/lib/utils imports in shared/ui/* → @/shared/lib/utils
