+++ shared/ui/label.tsx
@@ -2,7 +2,7 @@
 
 import * as React from "react"
 
-import { cn } from "@/lib/utils"
+import { cn } from "@/shared/lib/utils"
 
 function Label({ className, ...props }: React.ComponentProps<"label">) {
   return (
